fix: Improve performance of `medications` queries
TPP provide a `MedicationDictionary` table which maps internal drug IDs
to standard dm+d codes. Where this dictionary is incomplete we want to
be able to supply custom entries. The facility to do so was added in:

 * https://github.com/opensafely-core/ehrql/pull/1348

Unfortunately the combination of the two joins and the COALESCE and
NULLIF expressions caused abysmal query performance.

We can test this using a simple query template:
```sql
SELECT COUNT(*) FROM (
  /* table definition */
) AS t
WHERE t.date >= '20220101' AND t.date <= '20221231' AND t.dmd_code IN (
  /* prednisolone codelist */
)
```

Using the old, single-join table definition given here the query takes
48 seconds:
https://github.com/opensafely-core/ehrql/blob/488a9933/ehrql/backends/tpp.py#L106-L112

Using the two-joins-and-coalesce approach the same query takes 22
minutes:
https://github.com/opensafely-core/ehrql/blob/a17cbeab/ehrql/backends/tpp.py#L114-L122

The approach we take in this PR is to dynamically construct a single,
combined dictionary table with NULL entries removed and then do a single
join against that. It's not obvious that should perform better but it
turns out that it does: the same query using this approach takes just
over 1 minute.

Prednisolone codelist, for reference:
https://github.com/opensafely-core/ehrql/blob/a17cbeab/tests/acceptance/external_studies/openprompt-long-covid-vaccines/codelists/opensafely-asthma-oral-prednisolone-medication.csv