fix: Remove join between `Appointment` and `DataDictionary`
This join (added in #1413) looks innocuous but it caused a catastrophic
performance regression: queries which previously took 15 minutes were
taking 11 hours.

The problem is compounded by the presence of the DISTINCT clause (which
we hope is only temporary, see #1053) which seems to increase query
times by 4-5x. However this is dwarfed by the join which increases query
times by another 40-50x on top of that.

Replacing the join with a hardcoded case expression is only marginally
slower than not including the `status` field at all so we do that
instead.