Move example from `reference/language.md`
To standalone files.

Slightly irritatingly, the way we have the imports and the snippet
markers conflicts with how Black and Ruff prefer to have things. For
now, I've just gone with how the autoformat works, but we may want to
revisit this in future.

We could workaround this by using line numbers in the snippet selection,
instead of marker comments, although that's less robust. If someone
edits the example and the line numbers change, then the selection that
appears in the documentation changes.