docs: Document `patients` table
We don't have a mechanism for adding a backend-specific docstring to a
core table, so if we want to link to the paper, which is about the
OpenSAFELY-TPP backend, then we have to link to it from every backend.
Nevertheless, it's reasonable to assume that a researcher might want to
read the paper, regardless of the backend.

The information about orphan records comes from [this Slack thread][1].

Closes #1381

[1]: https://bennettoxford.slack.com/archives/C02HJTL065A/p1689327065988359