fix: Increase number of examples used in generative tests comprehensive check
The previous number of examples was mostly allowing this comprehensive
check to pass, but it was quite close to the minimum needed, so would
fail on the occasional local run. Increasing it to 250 adds around 1.5s
to the time the tests take to run, but should make the test more robust.