Reorder the navigation panel
ehrQL's documentation draws heavily on the [Diátaxis][1] approach. The
documentation for Diátaxis is ordered Tutorials, How-to guides,
Reference, and Explanation (i.e. clockwise, from the north-west
quadrant). This isn't an accident: it gives prominence to the
horizontal, Serve our study/Serve our work, axis that we know is also
important to our users. So, we reorder the navigation panel to do the
same.

[1]: https://diataxis.fr/