docs: Replace bulleted list with paragraph
Whilst the structure of the paragraph follows the bulleted list, the
text is taken from GLOSSARY.md.

Anecdotally, we know that the documentation is "written documents
heavy". To reduce the weight, I've not included information about
columns. Similarly, and also because not all data sources exist in ehrQL
at present, I've not linked to the Data Sources section of the
documentation.