docs: Make info box a section
This makes the info box that describes the relationship between ehrQL
and cohort-extractor a section. This is because info boxes are highly
visually salient; we don't want that level of visual saliency, here.

Although the diff might be hard to unpick, I've not changed the text in
the info box; I've simply outdented it.