docs: Move example
The example is the only content on the introduction-to-concepts page
that isn't also on the landing page. Consequently, if we move the
example, then we can remove the introduction-to-concepts page.

The (small) exception is that the introduction-to-concepts page
describes ehrQL as a *Python* programming language. Whilst this is true,
I think there's a strong case for not emphasising Python.