fix: Remove IterWrapper class
This was previously used as an opaque wrapper around iterables which
couldn't naturally be used in the query model because they weren't
hashable. However, this introduced indeterminacy into the query model
because the hashes of these IterWrapper object would depend on their
location in memory, and thus the hashes of everything which referenced
them would become indeterminate. And, because objects' hashes affect the
iteration order when you pull them out of a set, identical queries would
be processed slightly differently each time you ran them.

In general, the query model is much easier to reason about and better
behaved if it is only allowed to contain value objects.