Check Docker version supports Trino image
For boring reasons I ended up with quite an old version of Docker
installed locally and this caused the Trino image to error out with a
"Java is not installed" message, which was a bit confusing as it doesn't
look on the face of it like a Docker version issue.

This adds a Docker version check to the Trino fixture to save any future
confusion.