Reduce generative test example count
My plan was to bisect my way to a value that actually completes. Halving
from 10,000 to 5,000 allowed the tests to run correctly for long enough
that they found an actual novel bug. Running with 7,500 died with the
same error as before, so I've decided to stop at 5,000.

This obviously isn't ideal but, given that the tests we run are
different each time, I think it's strictly better to be doing some
generative testing every night even if it's less than we'd like.