docs: Describe an OpenSAFELY command
setting-up already describes an OpenSAFELY command. As at least one user
as asked about what the parts of an OpenSAFELY command mean, we describe
another here. We intentionally don't reference from one to the other
(and vice versa) as the tips are collapsed by default and a user may not
have read them.