Separate example from backend-tables page
This does reuse the tutorial example to avoid having to construct
another minimal example purely for a single import statement.

There is more overhead in the minimal example, because we need a full
working dataset definition. (It seems a possibly better approach to only use
`opensafely` to run dataset definition-related code.)

This uses output from the dummy data.