docs: Don't reference the project pipeline
When reading these tips, @CLStables said "I might not know what the
project pipeline is". I don't think that a user needs to know, at this
stage, what the project pipeline is. However, I do think that the user
needs to know that there are such things as OpenSAFELY actions and that
these can be executed independently.