fix: Accept subset of categories when reading Arrow files
If we have a column in a file where the available categories are A, B
and C it's perfectly valid for the file to contain only A and B.
Therefore we shouldn't enforce that the categorical dictionary contains
C as well.

There's a wider question around how we help users make sure their code
can cope with all the values the real data might throw at it when those
values aren't represented in the dummy data. But enforcing an exact
match for Arrow categoricals isn't the way to achieve this.