fix: Handle case where we happen to generate completely empty batch
I think I had implicitly assumed this was too unlikely to worry about,
but it turns out (thanks Hypothesis) that if you craft an inline table
with just the right patient IDs you can make this happen reliably.

The solution is to explicitly pass the SQLAlchemy metadata through,
rather than relying on there always being at least one object to carry
it.

It's tricky to test for this as it relies on generating just the right
pseudo-random sequence. The example Hypothesis came up with reproduces
reliably at the moment, but I think it would be sensitive to minor
changes in the code and thus wouldn't be a great test to have around.