Fiddle generative test settings
An effect of running multiple small batches is that we're more likely to
randomly hit a batch which generates a larger than usual number of
ignored errors. So we increase the threshold at which we complain about
these.

We also reduce the batch size and maximum query depth to get batches to
run in a more sensible time. (I've not seen any bugs exposed by very
deeply nested queries that wouldn't also be exposed by just a few layers
of nested operations.)