docs: Make example data one-indexed
Patient IDs are saved in CSV files, so I've simply incremented each by
one. Row IDs are generated, so I've modified a couple of __repr__
methods in the in-memory database.

Careful readers will notice that patient and row IDs in representations
of patient tables are left-aligned, whilst these IDs in representations
of event tables are right-aligned. By default, Python right-aligns
numbers [1], so I'd argue that these IDs aren't represented correctly in
patient tables. However, correcting this in `PatientTable.__repr__`
proved to be more trouble than it was worth.

[1]: https://docs.python.org/3/library/string.html#format-specification-mini-language

Closes #1620