Bump ignored error limit again
Running lots of small batches massively increases the chances that we
hit this.

Probably what we should do is write the stats to a file for each batch
and then some check that runs after all the batches have completed
which checks the cumulative error rate. But increasing the threshold
feels like a lot less work to me.