fix: Replace negative IMD values with NULLs
Although the issue mentions negative IMD values, the only negative IMD
values are -1s. So, we can use NULLIF to replace -1s with NULLs.

I've checked the performance impact of this change by comparing the
execution time of the new query (NULLIF) to the old query.

The first template query counted the number of rows where imd_rounded
was GTE zero. In this case, there was an 88ms difference in elapsed time
between new and old, favouring old.

The second template query counted the number of rows where imd_rounded
was NULL (new query) or was -1 (old query). In this case, there was a
33ms difference in elapsed time between new and old, favouring new.

Fixes #1548