Install ehrql package in development environment
When running code in a subprocess we can no longer rely on the ehrql
package always happening to be importable from the current working
directory. Instead we need to make it explicitly available to Python.

This also has the advantage that the command line entrypoint is
installed in the local development environment so you can run `ehrql
<command>` rather than `python -m ehrql <command>`.