fix: Always validate categorical data in user-provided arrow dummy dataset
When a user provides their own dummy dataset, we read it in and then
write it out again.  If a dataset variable is categorical, and if the
dataset is written out to an arrow file, an error will be thrown by
pyarrow if the corresponding column contains unexpected values.

As such, when we read in a user's dummy dataset from an arrow file, we
need to validate that the values in a categorical column are in the set
of expected values, even if pyarrow does not think that the column is
categorical.