Don't recompile requirements when installing
Because we use timestamps to test for modifications this ends up getting
run all the time when rebasing or hopping between branches. Installing
requirements only involves local operations and is generally very fast
(especially when it's a no-op) but _compiling_ requirements requires
talking to PyPI and can be intolerably slow.

And I think in any case this is the wrong behaviour: re-compiling
requirements should be an explicit operation.

Further discussion here:
https://bennettoxford.slack.com/archives/C63UXGB8E/p1693556976925389