fix: Upgrade to Pledge v1.9
Pledge is now bundled as part of the "Cosmos" suite of software:
https://github.com/jart/cosmopolitan/releases/tag/3.0.1

The changes since v1.8 are a _bit_ cryptic but look worth pulling in:
https://github.com/jart/cosmopolitan/commits/55d5d6bbc/tool/build/pledge.c

Note the binary is significantly larger because this is the
cross-platform APE build rather than the Linux only one. This is a bit
pointless for our purposes as the features we care about are Linux only.
But it's still only 540K and there's a huge advantage to just using the
official build.

This introduces a slight wrinkle in that the APE format does some
beautifully terrible trickery to work across different platforms which
means that, by default, you need to execute them via a shell (gory
details at https://justine.lol/apeloader/). To avoid this we have to
"assimilate" the binary to a native format. This is all wrapped up in a
Just command which downloads, extracts, and assimilates the binary:

    just download-pledge