Let both Backend and QueryEngine influence column types
This involves some unpleasant back and forth between the Backend and
QueryEngine classes ("I'll call your hook if you call mine") because,
under different circumstances, either might need to be responsible for
setting the column types here.

What I think this shows is that we haven't succeeded in factoring out the
responsibilities between the two classes correctly here. But we've
decided to punt on this question for now as long as we've got something
that does what we need.