Add `ons_deaths` PatientFrame table
In contrast to raw.ons_deaths, ons_deaths only contains one row per
patient and is therefore a PatientFrame. If the dates are different
ons_deaths always contains only the first entry. In cases where the
date is the same the underlying cause of death (icd10u) is arranged
alphabetially and the first one will be selected. In cases where the
date of death and the underlying cause of death are the same, the
selection will be nondeterministic.