Update use of `ons_deaths` in acceptance tests
This updates the code use of ons_death for external studies used in our
acceptance tests to match the new, intended behaviour.
Because ons_deaths is now a PatientFrame (with one row per patient
only) it is no longer needed (or possible) to perform the .sort_by()
and .first_for_patient() operations.