fix: Fx bug in MSSQL horizontal aggregation workaround
We work around the absence of `GREATEST/LEAST` support in our version of
MSSQL using a table value constructor and a scalar select. This required
a hacky workaround to get SQLAlchemy's ScalarSelect object to "remember"
whch tables it was selected from. However this hack turned out to be
insufficient: it handled the `_from_objects` property but not the
`get_children()` method.

Adding support for large codelists in #1710 exposed the problem here as
it became possible to construct aggregations which would "lose" the
reference to the codelist table and therefore never create it.

Fortunately, Hypothesis found this issue for us before it affected any
users. Nice one.

The workaround we now use (making our own ScalarSelect subclass with the
desired behaviour) is still a bit of a hack, but slightly less so; and
it solves the problem in question.

Closes #1724