feat: Add time limit for generating measures
This add a new class MeasureTimer to estimate the time it would take
to iterate through all measures starting from the 12th iteration
(informed guess to avoid stopping jobs that had a slow start) and
raises an Exception (MeasuresTimeout) if a specified time limit
(3 days by default, also just a guess) is exceeded.
The counter increases by 1 with each new interval period.
This allows to count the actual number of queries without having to look
into get_results().

The first iteration in the test mocks a slow iteration that would
exceed the default time limit, but we only test this from 12 iterations
onwards. The following 11 iterations (500 sec) bring the projected
time down so that the 12th iteration does not raise the MeasuresTimeout.
The 13th iteration in the test is considerably longer and pushes
the projected time over the limit and raises the MeasuresTimeout.