fix: Dynamically re-collate columns where needed
The reason this matters is that when we create a temporary table
containing a list of code which we want to use to query this column we
need the collations to match. We work on the assumption that a given
coding system always uses the same collation within the TPP database but
this turns out not to be the case (presumably because the tables in
question are externally provided).

To workaround this we re-collate the columns when querying them. This
isn't going to be good for performance but I think the tables in
question are small enough (i.e. not in the billions of rows) where we
can get away with this.