Explicitly collate string literals and pseudo identifiers
This ought not to be necessary because the collation we're targetting
here is the TPP server default. But not having it makes testing
difficult because it isn't test server default. And we can't change the
test server default because we need a case-sensitive collation to run
the tests against.

I'm hopeful that, while they make the SQL more cluttered, these changes
won't have any material impact on execution time.