Add disclosure control for measures
This adds disclosure control -- suppressing small numbers (less than or
equal to seven) and then rounding numbers (to the nearest five) -- to
the measures framework. The OpenSAFELY approach to disclosure control is
documented in "Updated disclosure control guidance" [1].

Disclosure control is enabled by default. It is disabled by calling:

```
measures.configure_disclosure_control(enabled=False)
```

The OpenSAFELY docs [2] suggest suppressing small numbers by replacing
them with `"[REDACTED]"` (i.e. a string). However, doing so adds
complexity to the implementation, because some output formats have typed
columns. I considered replacing small numbers with `None`, but this does
the same: The complexity shifts from the output format to the
calculation of the ratio.

Replacing small numbers with zeros doesn't add complexity to the
implementation. As well as being of the same type, the calculation of
the ratio before disclosure control is the same as the calculation of
the ratio after disclosure control.

Consequently, small numbers are replaced with zeros.

[1]: https://www.opensafely.org/updated-output-checking-processes/
[2]: https://docs.opensafely.org/releasing-files/#redacting-counts-less-than-or-equal-to-7