Disable disclosure control in example
I think this section is easier for a reader to understand when the
example disables disclosure control. They can picture a number of
patients in the denominator group and a number of patients in the
numerator group, without also having to step through how those numbers
would be suppressed/rounded.