fix: `UnboundLocalError` for `elapsed_time` variable
This bug was initially not exposed by our test because we did not
have any groups in the test measure. The main fix is changing the
variable elapsed_time to an instance variable self.elapsed_time.
Other small changes include: (1) start taking the time from the
12th iteration onwards (this change is also reflected in the test)
and (2) change the name of the first argument of the classmethod
from_grouped() to match the convention cls.