feat: Update patients and clinical_events tables in EMIS backend
In EMIS, the column that joins all patients is registration_id, not
patient_id. The patients table contains a very small number of
duplicates, and this uses the same strategy to deduplicate them as
cohort-extractor i.e. we just ignore them altogether.

The clinical_events table is updated to use the correct EMIS
underlying table and column names, as well as casting to the types
that the table in `tables.beta.core` expects them to be.