Correctly determine tables used in population expression
We used to rely on SQLAlchemy to do this for us by building a select
query using the expression and calling `get_final_froms()` on it.
However this doesn't work for tables which appear in the expression as
part of a correlated subquery – by design, such tables are excluded from
the FROM clause.

Now we unpack the expression ourselves to build a list of referenced
tables.