Change ehrQL's empty list semantics to match SQL's
As discussed and agreed in the Data Team meeting, we can describe
ehrQL's approach to NULLs as: mirror SQL's semantics except in
aggregations we were remove NULLs rather than treating them as unknown
values.

This change simplifies both the code to generate the SQL and the SQL
it generates. Given that the most likely use of these expressions is as
the argument to a `where()` method in which NULL and False have the same
behaviour, our decision here is unlikely to make any practical
difference.