feat: Add EMIS vaccinations table
Adds the EMIS vaccinations table. In cohort-extractor, the method to
query vaccinations also allowed querying by product code, but it did
so by inner joining the medications table on the immunisation table.
That gives us all medications for patients who also had a vaccination
procedure, which isn't the right thing to do here - a lot of those
medications will not be vaccination-related. In order to get the
same behaviour as cohort-extractor, users can query this vaccinations
table with a codelist for vaccination procedure codes, and the
medications table with a codelist for vaccination product codes.