chore: Exclude `/includes/` directory
(Yes, that doesn't sound like a contradiction at all.)

Files in the top-level `/includes/` directory are actually included by
the snippet support in pymdown-extensions, instead of directly included.

This exclusion is to:

* prevent the `includes/` directory from being included as built HTML
  pages themselves (which is in addition to their use as snippets in
  other built HTML pages)
* allows suppressing of MkDocs warnings from these pages for `mkdocs build` and
  `mkdocs serve --clean`. (These warnings are because the relative links in the
  `includes/` source don't point to their intended actual links in the
  source directory, but they do when included as snippets.)

The preceding `/` in `/includes/` is to exclude only the top-level
`/includes/` directory to keep the exclusion rule specific; the option
uses a `.gitignore` pattern format.

See https://www.mkdocs.org/user-guide/configuration/#exclude_docs