chore: Use correct relative links
Some of the existing links work in the ehrQL docs build,
but not the main OpenSAFELY documentation site.

This is because:

* some of the existing links are either absolute links to the URL root,
  or relative paths that go the URL root, before descending
* the ehrQL docs build preview has all the content at the
  URL root, and not, like the OpenSAFELY documentation site, in
  `/ehrql/`, so there's a mismatch of the depth of URLs there

It would be good to make the two sites consistent: that is, make the
ehrQL preview actually have URLs like: `/ehrql/how-to/…` instead of
`/how-to/…` — it would remove the possibility for this kind of broken
link.