test: Add code to test measures
For now, use a crude search of the source text to distinguish between
different kinds of ehrQL definition code.

The current `ehrql` syntax label used in the source does not
discriminate between different kinds of ehrQL definitions.

We could make authors label the code blocks accordingly, either:

* with a specific label `ehrql-dataset` and `ehrql-measures`, for example
* or, repurposing
  [HTML attributes](https://facelessuser.github.io/pymdown-extensions/extensions/superfences/#injecting-classes-ids-and-attributes)
  on the examples; for example, `{.ehrql .dataset}` and `{.ehrql .measures}`

However, we currently support Python source files too; and these would
need some way of being labelled (which could be via filename).

For now, it's easier to just search for `create_dataset()` or
`create_measures()`. We could use `ast` to extract function calls, but
not sure it's worth the slight extra complexity: it means that when you
create the `EhrqlExample`, it's possible that parsing fails because the
example is invalid Python, and you need to handle those exceptions. It's
a nice property that we handle the example source as some string that
may/may not be valid until the actual test runs.