Check ingored test rate in CI not in overnight tests
The purpose of this check is to make sure that we don't muck up the
"ingore broken examples" mechanism such that we end up not actually
executing any generative tests, although we think we are.

This is best done in CI, where things are deterministic and where we can
catch before merging, rather than in the overnight tests. We run so many
batches of examples in the overnight tests that just by chance we can
end up hitting a lot of ignored examples every now and then.