chore: Use `--clean` as MkDocs default
This switch is slightly overloaded in MkDocs, and not only removes old
files from the `site_dir` but also enforces the `exclude_docs` setting
in `mkdocs.yml`.

`--clean` is the default for `mkdocs build` but not `mkdocs serve`. It's
included here for `mkdocs build` to make the option explicit, and to not
raise the question of why `serve` has `--clean`, but `build` does not.