fix: Support sorting by patient series
The in-memory and MSSQL engines need updating here but the others handle
this already. Given that these operations are no-ops we could filter
them out earlier, but I have more confidence in the correctness of our
implementation if Hypothesis can compare the behaviour of engines which
filter these operations with those that handle them directly.