fix: Fix MSSQL sort-by-constant bug
This is a follow-up to the attempted fix in #1869. I had naively assumed
that the set of all constant expressions was a subset of all
patient-level expressions and therefore, as both are no-ops when used as
sort conditions, we could just exclude all patient-level expressions –
which we already have code to identify.

However, as Hypothesis helpfully [pointed out][1], there are event-level
expressions which evaluate as constant e.g. `event_series.is_in([])`,
which is always false.

We now add a mechanism for explicitly identifying constant expressions.

Hopefully actually closes #1864

[1]: https://github.com/opensafely-core/ehrql/issues/1864#issuecomment-1876693379