feat: Add `validate` methods for constraints
Each validate method checks the constraint and returns a boolean with
the result. All validate methods apart from the NotNull constraint
return True if the value to be ckecked is None. This allows us to only
check values that are present. We cant check the Unique constraint with
a single number so this currently always returns True.