Rename user-facing error
We want to avoid leaking unnecessary internal details into our error
messages so as to minimise the amount of syntactic noise users have to
deal with.

This changes exceptions from:

    ehrql.query_language.InvalidOperationError: ...

to:

    ehrql.Error: ...

There's a bit of skullduggery with the `__module__` attribute to make
this all work nicely, but I think it's worth it.