feat: Add `covid_therapeutics_raw` table
This adds a mostly raw (rare?) view on the COVID Therapeutics dataset.
Things which could be straightforwardly tidied up (e.g. converting
datetimes to dates) but anything more exotic has been left as is.

Values that have been converted have been checked against the real data
to confirm that the conversions are valid.

Documentation has been pulled in from the old Cohort Extractor docs:
https://docs.opensafely.org/data-sources/therapeutics/

And also from the very useful Short Data Report (linked in the code).

There's a further question as to how we properly tidy this, remove
duplicates etc; but I thought it was best to start with this.

Closes #1896