Increase examples for comprehensiveness check
Following changes in Hypothesis we now need to generate more examples to
get full coverage of the query model. This increases the time needed to
run the check from just over 5 seconds to just over 11 seconds, which is
annoying but not the end of the world.

The changes in question are presumably those described here:
https://hypothesis.readthedocs.io/en/latest/changes.html#v6-98-1

The full set of relevant commits is here:
https://github.com/HypothesisWorks/hypothesis/compare/hypothesis-python-6.96.2...hypothesis-python-6.98.1