fix: Fix edge case in constructing population table
Previously, the code made an implicit assumption that if the SQL
expression for a population definition referenced only a single table
then that table must be one-row-per-patient.

This happened to be true until we introduced the `CombineAsSet`
aggregtation, but it was never a sound assumption.

Naturally, its falsity was discovered by the gentests.

Closes #1918