Disable statement caching for SQLite dialect
This doesn't benefit us at all (it's designed to avoid the compilation
overhead when executing the same SQL over and over) and it leads to
warnings when we execute clause elements whose caching behaviour has not
be explicitly defined.

We're getting a warning here now because we've just updated to Trino
v0.328.0 which includes this PR:
 * https://github.com/trinodb/trino-python-client/pull/429

This defines a compiler for the `Lag` clause which (because it is not
restricted to a specific dialect) ends up being used by SQLite and
triggering the warning.