Bump Trino version
The previously specified version was chosen arbitrarily: it was just the
most recent version at the point when development started on the Trino
query engine.

However, this version contains the bug identified in #1715 and
Hypothesis keeps rediscovering this bug which devalues the generative
tests.

Eventually we wil need to resolve the question of exactly what EMIS are
running in production and ensure we're testing against that. But for now
there's no benefit to not using the latest version.