Enforce consistent docstrings for inherited tables
Docstrings are a bit of a wrinkle here: they obviously aren't
automatically inherited from the parent; we could set them dynamically
by fiddling about with the `__doc__` attribute but that would prevent
them showing up in VSCode, and it's quite a nice feature that they
display directly in the editor.

So instead we require them to be manually copy-pasted but we add a check
that they are consistent to avoid drift. We don't require that they are
identical because it seems reasonable to allow the child table to extend
the parent docstring, but they must share a common prefix.