Move dependency list out of `pyproject.toml`
They were moved in as part of #252 on the basis that this was "more
modern". However we don't – and don't intend to – publish ehrQL as a
package to PyPI and when we do build it as a package in our Docker image
we tell pip to ignore dependencies anyway. And for some reason that we
haven't got to the bottom of, Dependabot is unable to update our
production dependencies if we specify them in this way.

This moves us back to the `requirements.prod.in` pattern that we use
elsewhere and which Dependabot handles nicely.

Closes #964