Reject invalid `when|then` arguments immediately
The previous "lazy" approach would only attempt to convert the arguments
to query model objects when the `case()` expression was evaluated
meaning that errors would only be triggered at that point.

Performing the conversion eagerly means that we trigger the error right
at the relevant point in the code, which is more helpful. It also means
we can do away with the `_DictArg` nonsense.