Generate better test IDs for lambdas
By default, when using lambdas as test parameters, pytest generates
uninformative test names like:

    tests/unit/test_module.py::test_function[<lambda>1]

Getting the source code for a given lambda expression is achievable but
a bit fiddly so we borrow Hypothesis's function for doing it. This gives
us more helpful test names like:

    tests/unit/test_module.py::test_function[x + 100]

Which make it easier to identify the problematic example in the case of
failure.