fix: Reject duplicated conditions
Previously this what have passed without error but only the last
specified condition would have taken effect. This violates the
documented behaviour of `case()` where conditions are matched in the
order given and the _first_ match takes effect. We could fix this
behaviour easily enough, but it's presumably a mistake anyway so better
to alert the user immediately.