Update end-to-end test fixtures to pass T1OO filter
In a sense this isn't ideal because we're relying on particular features
of our T1OO logic and if this changes the tests may need updating again.
The alternative would be threading the appropriate `include_t1oo`
database connection parameter through to the test container.  However,
given the role of these tests as end-to-end tests of the system it seems
better to stick to default parameters and exercise the full T1OO code
path.