Update t1oo logic to reflect decision on deregistration window
We only include t1oo patients who are:
- still registered OR
- deregistered, but died before or on deregistration date

Due to delays in the flow of information around the system, it is
possible that patients may have a date of death recorded shortly AFTER
deregistration. However, we take the most cautious approach, and
only include patients if they died before or on deregistration date.