fix: Handle non-dataclass objects in graph renderer
The query graph renderer is only used internally, and then only really
for demonstration purposes, so it's not heavily tested. This fixes an
issue wherey query graph objects which are not dataclasses (e.g.
`Position` enums or `RowsReader` instances) would cause the renderer to
fall over. We do this by inverting the logic to explicitly check for
problematic primitive types.