fix: Enable `unveil` in user-supplied code sandbox
This tightens even further the restrictions placed on the sandboxed
process in which we run user-supplied code. Thanks to a recent upgrade
of the VM guest OS the `unveil` feature is now supported which provides
more fine-grained control over what paths are readable by the sandboxed
process. This means we can prevent the process reading its parent's
environment variables. There was never anything the process could _do_
with these variables (being denied all network access and filesystem
write access) but it's better if it can't read them at all.