Fix assimilation of `pledge` binary within Github Actions
Cosmopolitan/APE is dark magic and something has changed within the
Github Actions runtime which means that executing the APE binary
directly doesn't work and we need to do it via a shell. Once the binary
has been "assimilated" it can then becomes a native binary which can be
executed normally.