Add get_exit_status_for_exception methods to backends
Adds a methods by which a backend can return custom exit codes
for specific exceptions. This is backend specific, and will depend on
the exceptions that we know we need to handle for each individual
backend.

This method takes and exception, and may return an (exit_code,
log_message)
tuple for us to exit with, or may return None, in which case we'll
just raise it and continue as normal.

We add the same exception handling for the TPP backend as we have in
cohort-extractor
https://github.com/opensafely-core/cohort-extractor/blob/52588b9518c67632ddae98780d58078c947a720d/cohortextractor/cohortextractor.py#L922-L947