fix: Temporarily revert to denylist for T1OO
This reflects the new interim agreement with respect to T1OOs. This
change is only required while TPP update the logic for populating the
allowlist. Once that is complete we can switch back to using the
allowlist. For this reason I am leaving some of the end-to-end test
setup in place even though it is not strictly required while using the
denylist.

Reverts some of #2072