Remove structlog, just use stdlib logging
We were not using most of the features of structlog in ehrql log
outputs.  The two we were using were coloured output, which we think we
want to remove anyway, as it causes issue using logs in productions.
The other was k=v formatted extra values, which we only use in one
place.

stdlib logging can do the same job, and we just include the k=v tags
a bit more manually when needed.  This removes a dependency, which is
always nice, but also means that the ehrql sandbox can work without
*any* external dependencies, which is quite a nice property.

We may end up reintroducing coloured outputs again, but hopefully not by
embedding ANSI colour codes in log files.